HUGE PR WITH ALL THE WORK FOR THE PAST YEAR WITH NO ORDER

+++ src/server/api/routers/player.ts
@@ -1,31 +1,74 @@
+import meta from '@/server/api/meta';
 import {
+  authProcedure,
   clubAdminProcedure,
   protectedProcedure,
   publicProcedure,
 } from '@/server/api/trpc';
-import { deletePlayer, editPlayer } from '@/server/mutations/club-managing';
+import { clubsSelectSchema } from '@/server/db/zod/clubs';
+import {
+  playerAuthStatsSchema,
+  playerEditSchema,
+  playerFormSchema,
+  playersSelectSchema,
+  playerStatsSchema,
+} from '@/server/db/zod/players';
+import { playerToTournamentSchema } from '@/server/db/zod/tournaments';
+import { usersSelectMinimalSchema } from '@/server/db/zod/users';
+import {
+  createPlayer,
+  deletePlayer,
+  editPlayer,
+} from '@/server/mutations/club-managing';
 import {
   abortAffiliationRequest,
   acceptAffiliation,
+  affiliateUser,
+  cancelAffiliationByUser,
   rejectAffiliation,
   requestAffiliation,
 } from '@/server/mutations/player-affiliation';
 import getPlayer from '@/server/queries/get-player';
-import getPlayersLastTmts from '@/server/queries/get-players-last-tmts';
+import { getUserClubIds } from '@/server/queries/get-user-clubs';
+import {
+  getPlayerAuthStats,
+  getPlayerStats,
+  getPlayersTournaments,
+} from '@/server/queries/player';
+import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 
 export const playerRouter = {
   info: publicProcedure
+    .meta(meta.playerInfo)
     .input(z.object({ playerId: z.string() }))
+    .output(
+      playersSelectSchema
+        .extend({
+          user: usersSelectMinimalSchema.nullable(),
+          club: clubsSelectSchema,
+        })
+        .nullable(),
+    )
     .query(async (opts) => {
       const { input } = opts;
       return await getPlayer(input.playerId);
     }),
-  playersLastTournaments: publicProcedure
+  create: clubAdminProcedure
+    .meta(meta.playersCreate)
+    .input(playerFormSchema)
+    .output(playersSelectSchema)
+    .mutation(async (opts) => {
+      const { input: player } = opts;
+      return await createPlayer({ player });
+    }),
+  lastTournaments: publicProcedure
+    .meta(meta.playersLastTournaments)
     .input(z.object({ playerId: z.string() }))
+    .output(z.array(playerToTournamentSchema))
     .query(async (opts) => {
       const { input } = opts;
-      return await getPlayersLastTmts(input.playerId);
+      return await getPlayersTournaments(input.playerId);
     }),
   affiliation: {
     request: protectedProcedure
@@ -78,33 +121,75 @@ export const playerRouter = {
         const { input } = opts;
         await abortAffiliationRequest(input);
       }),
+    affiliateAuth: clubAdminProcedure
+      .input(
+        z.object({
+          playerId: z.string(),
+        }),
+      )
+      .mutation(async (opts) => {
+        const {
+          input,
+          ctx: { user, clubId },
+        } = opts;
+        const { playerId } = input;
+        await affiliateUser({ playerId, user, clubId });
+      }),
+    cancelByUser: protectedProcedure
+      .input(
+        z.object({
+          playerId: z.string(),
+        }),
+      )
+      .mutation(async (opts) => {
+        const { input } = opts;
+        await cancelAffiliationByUser({ userId: opts.ctx.user.id, ...input });
+      }),
   },
-  delete: clubAdminProcedure
+  delete: protectedProcedure
+    .meta(meta.playersDelete)
     .input(
       z.object({
-        clubId: z.string(),
         playerId: z.string(),
-        userId: z.string(),
       }),
     )
-    .mutation(async (opts) => {
-      const { input } = opts;
+    .output(z.void())
+    .mutation(async ({ input, ctx }) => {
+      const clubs = await getUserClubIds({ userId: ctx.user.id });
+      const player = await getPlayer(input.playerId);
+      const isAdmin = Object.keys(clubs).find(
+        (clubId) => clubId === player.clubId,
+      );
+      if (!isAdmin) throw new TRPCError({ code: 'UNAUTHORIZED' });
       await deletePlayer(input);
     }),
   edit: clubAdminProcedure
-    .input(
-      z.object({
-        clubId: z.string(),
-        values: z.object({
-          id: z.string(),
-          nickname: z.string(),
-          realname: z.string().nullable(),
-          rating: z.number(),
-        }),
-      }),
-    )
+    .meta(meta.playersEdit)
+    .input(playerEditSchema)
+    .output(playersSelectSchema)
     .mutation(async (opts) => {
       const { input } = opts;
-      await editPlayer(input);
+      return await editPlayer({ values: input, user: opts.ctx.user });
     }),
+  stats: {
+    public: publicProcedure
+      .meta(meta.playersPublicStats)
+      .input(z.object({ playerId: z.string() }))
+      .output(playerStatsSchema)
+      .query(async (opts) => {
+        const { input } = opts;
+        return await getPlayerStats(input.playerId);
+      }),
+    auth: authProcedure
+      .meta(meta.playersAuthStats)
+      .input(z.object({ playerId: z.string() }))
+      .output(playerAuthStatsSchema.nullable())
+      .query(async ({ input, ctx }) => {
+        if (!ctx.user) return null;
+        return await getPlayerAuthStats({
+          playerId: input.playerId,
+          userId: ctx.user.id,
+        });
+      }),
+  },
 };
